requirements-py313.txt
numpy>=2.1.1
pydicom>=2.4.4
nibabel>=5.2.1
matplotlib>=3.8.4
tqdm>=4.66.4
fpdf2==2.7.9
Pillow==10.4.0
SimpleITK>=2.3.1

